+++ DoAn-2/MenuTab/frmNhapKho.cs
@@ -155,13 +155,14 @@ namespace restaurant_management.MenuTab
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
-            if (MessageBox.Show("Bạn chắc chắn muốn xoá?", "Xác nhận", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
-            {
-                if (PhieuNhapDAO.Instance.DeletePhieuNhap(int.Parse(txtMaPhieuNhap.Text)))
-                    MessageBox.Show("Xoá thành công!");
-                SetBtnEdit_Off();
-                LoadData();
-            }
+            if (!string.IsNullOrEmpty(txtMaPhieuNhap.Text))
+                if (MessageBox.Show("Bạn chắc chắn muốn xoá?", "Xác nhận", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                {
+                    if (PhieuNhapDAO.Instance.DeletePhieuNhap(int.Parse(txtMaPhieuNhap.Text)))
+                        MessageBox.Show("Xoá thành công!");
+                    SetBtnEdit_Off();
+                    LoadData();
+                }
         }
 
         private void btnThem_Click(object sender, EventArgs e)
