+++ DoAn-2/MenuTab/frmVoucher.cs
@@ -111,13 +111,14 @@ namespace restaurant_management.MenuTab
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
-            if (MessageBox.Show("Bạn chắc chắn muốn xoá?", "Xác nhận", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
-            {
-                if (VoucherDAO.Instance.DeleteVoucher(int.Parse(txtMaVoucher.Text)))
-                    MessageBox.Show("Xoá thành công!");
-                SetBtnEdit_Off();
-                LoadData();
-            }
+            if (!string.IsNullOrEmpty(txtMaVoucher.Text))
+                if (MessageBox.Show("Bạn chắc chắn muốn xoá?", "Xác nhận", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                {
+                    if (VoucherDAO.Instance.DeleteVoucher(int.Parse(txtMaVoucher.Text)))
+                        MessageBox.Show("Xoá thành công!");
+                    SetBtnEdit_Off();
+                    LoadData();
+                }
         }
 
         private void dtgvKhachHang_SortStringChanged(object sender, EventArgs e)
